pyproject.toml
[project]
name = "mustage-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "curl-cffi>=0.10.0",
    "fastapi[all]>=0.115.12",
    "loguru>=0.7.3",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "selectolax>=0.3.28",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
]
